+++ src/main/coerce.c
@@ -126,7 +126,7 @@ IntegerFromReal(double x, int *warn)
 	*warn |= WARN_NA;
 	return NA_INTEGER;
     }
-    return x;
+    return (int) x;
 }
 
 int attribute_hidden
@@ -140,7 +140,7 @@ IntegerFromComplex(Rcomplex x, int *warn)
     }
     if (x.i != 0)
 	*warn |= WARN_IMAG;
-    return x.r;
+    return (int) x.r;
 }
